apps/api/requirements.txt
# CohortLens API - Dependencies (Python 3.9+)

# Data processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Machine learning
scikit-learn>=1.3.0,<2.0.0
shap>=0.43.0,<0.45.0

# RAG (optional)
openai>=1.0.0

# Web3 / IPFS
requests>=2.31.0

# SaaS / Payments
stripe>=7.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Data validation
pandera>=0.17.0,<0.18.0

# Configuration
pyyaml>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# API & Web
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
strawberry-graphql[fastapi]>=0.200.0,<1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# CLI
click>=8.1.0,<9.0.0

# Reporting
jinja2>=3.1.0,<4.0.0
weasyprint>=60.0; sys_platform != "win32"
tabulate>=0.9.0,<0.10.0

# Statistical tests (drift)
scipy>=1.11.0,<2.0.0

# Utilities
joblib>=1.3.0,<2.0.0
tqdm>=4.66.0,<5.0.0

# Database (Neon DB)
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0

# Dashboard / charts
plotly>=5.18.0
streamlit>=1.28.0,<2.0.0
